src/main/java/eu/dissco/core/translator/terms/specimen/chronometric/ChronometricAgeConversionProtocol.java
package eu.dissco.core.translator.terms.specimen.chronometric;

import com.fasterxml.jackson.databind.JsonNode;
import eu.dissco.core.translator.terms.Term;
import java.util.List;

public class ChronometricAgeConversionProtocol extends Term {

  public static final String TERM = "chrono:chronometricAgeConversionProtocol";

  private final List<String> dwcaTerms = List.of(TERM);

  @Override
  public String retrieveFromDWCA(JsonNode unit) {
    return super.searchJsonForTerm(unit, dwcaTerms);
  }

  @Override
  public String getTerm() {
    return TERM;
  }
}
